ingest/stock/requirements.txt
autopep8==1.4
certifi==2018.8.24
chardet==3.0.4
idna==2.7
kafka-python==1.4.3
numpy==1.15.2
pandas==0.23.4
pycodestyle==2.4.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
urllib3==1.23
